Unfinished refactoring attempt: isomorphic models & fields
The sad story of this commit is recorded in more detail elsewhere, but
the long and short of it is: I tried to increase the ability to reuse
Concord constructs within conditions, actions, etc by creating basic
field and model types that could be re-used, but it's now two months
later and I feel like if this was a good approach it would have worked
by now.

The code in this branch is actually pretty solid and possibly worth
re-using later. Further developing it just can't be the priority
right now.